fix: prevent overflow in accessibility route cumulative distance
Dijkstra's cumulative distance in AccessibilityGraph was accumulated as
plain int, so a multi-hop route could silently wrap past
Integer.MAX_VALUE and corrupt shortest-path comparisons even though
every individual connection distance is a valid int. bestDistance,
QueueEntry, and GraphPath.totalDistanceInMetres now use long for the
running/reported total while individual edge distances stay int.
reconstructPath also gains a visited-set cycle guard so a corrupted
predecessor chain fails fast instead of looping indefinitely.

+++ src/main/java/seedu/unienable/logic/graph/GraphPath.java
@@ -6,10 +6,14 @@ import java.util.List;
  * An immutable shortest-path result from {@link AccessibilityGraph#getShortestPath}: the ordered
  * facility names visited from source to destination (inclusive of both), and the total distance
  * in metres.
+ *
+ * <p>The total is held as {@code long} rather than {@code int}: each individual connection
+ * distance is a valid {@code int}, but a multi-hop route's cumulative distance can exceed {@code
+ * Integer.MAX_VALUE}, so only the running total needs the wider type.
  */
 public final class GraphPath {
     private final List<String> facilityNames;
-    private final int totalDistanceInMetres;
+    private final long totalDistanceInMetres;
 
     /**
      * Creates a GraphPath.
@@ -17,7 +21,7 @@ public final class GraphPath {
      * @param facilityNames the facility names visited, in travel order, source to destination
      * @param totalDistanceInMetres the path's total distance in metres
      */
-    public GraphPath(List<String> facilityNames, int totalDistanceInMetres) {
+    public GraphPath(List<String> facilityNames, long totalDistanceInMetres) {
         this.facilityNames = List.copyOf(facilityNames);
         this.totalDistanceInMetres = totalDistanceInMetres;
     }
@@ -28,7 +32,7 @@ public final class GraphPath {
     }
 
     /** Returns the path's total distance in metres. */
-    public int getTotalDistanceInMetres() {
+    public long getTotalDistanceInMetres() {
         return totalDistanceInMetres;
     }
 }
